Fix pyexiv2 plugin to conform to refactoring
Also rename plugin from py3exiv2 to pyexiv2 to prevent confusion.
